lang_wheel: implement io
*wipes sweat* This was hard.

Tried to follow the guidelines in [1] (read: copypaste and then adapt)
but their version won't work for me, had to change double quotes for
backticks in the definite clause grammar for input.

And that forced me to use a separate rule for output because the one I
initially wrote only plays well with double quotes. It's most probably
an issue with types (character/s, character code, string) but it got too
complicated for me to give a shit.

In any case, this version just works. :rocket: Meaning, I can just run main and
the program will interpret the calendar and spit out the next day's
language according to constraints. :fire:

On character/string types, I found [2] very helpful when writing the
output part, and it may help clear the fog I have with the input DCG.

[1] https://www.metalevel.at/prolog/dcg
[2] https://stackoverflow.com/questions/27576261/convert-a-char-list-to-a-string-in-prolog