lang_wheel: make the program more script-like
It runs main on init and halts (exit) at the end of main. So it can be
run directly from the shell (or other scripts) without having to exit
the swi-prolog repl manually.

Taken from here:
https://stackoverflow.com/questions/25924247/running-prolog-as-a-script

Caveats:
- Now it's impossible to test things, cause the program halts even when
you load it from the repl (with consult). But I just need to remove the
initialization directive and then I can test the inner rules.
- I noticed that the program is pretty stealthy, appends a new line to
lang-calendar.txt but reports nothing to the shell...